Move test files closer to their counterparts
