+++ global.d.ts
@@ -42,3 +42,23 @@ declare global {
         }
     }
 }
+
+// Extend Jest's expect with custom matchers defined
+// and injected from @dd/tests/src/_jest/setupAfterEnv.ts
+interface CustomMatchers<R> {
+    toBeWithinRange(floor: number, ceiling: number): R;
+    toRepeatStringTimes(st: string | RegExp, occurences: number | [number, number]): R;
+}
+
+interface NonCustomMatchers {
+    toBeWithinRange(floor: number, ceiling: number): number;
+    toRepeatStringTimes(st: string | RegExp, occurences: number | [number, number]): string;
+}
+declare global {
+    namespace jest {
+        interface Expect extends NonCustomMatchers {}
+        interface Matchers<R> extends CustomMatchers<R> {}
+        interface InverseAsymmetricMatchers extends NonCustomMatchers {}
+        interface AsymmetricMatchers extends NonCustomMatchers {}
+    }
+}
