+++ packages/core/src/helpers/paths.test.ts
