+++ packages/core/src/helpers/strings.test.ts
