+++ packages/plugins/analytics/src/index.test.ts
