+++ packages/plugins/git/src/index.test.ts
@@ -6,9 +6,13 @@ import type { Options, RepositoryData } from '@dd/core/types';
 import { uploadSourcemaps } from '@dd/error-tracking-plugin/sourcemaps/index';
 import { getRepositoryData } from '@dd/internal-git-plugin/helpers';
 import { TrackedFilesMatcher } from '@dd/internal-git-plugin/trackedFilesMatcher';
-import { API_PATH, FAKE_URL, defaultPluginOptions } from '@dd/tests/_jest/helpers/mocks';
+import {
+    API_PATH,
+    FAKE_URL,
+    defaultPluginOptions,
+    getSourcemapsConfiguration,
+} from '@dd/tests/_jest/helpers/mocks';
 import { BUNDLERS, runBundlers } from '@dd/tests/_jest/helpers/runBundlers';
-import { getSourcemapsConfiguration } from '@dd/tests/unit/plugins/error-tracking/testHelpers';
 import nock from 'nock';
 
 jest.mock('@dd/internal-git-plugin/helpers', () => {
