+++ packages/plugins/telemetry/src/common/output/files.test.ts
@@ -4,8 +4,7 @@
 
 import { outputFiles } from '@dd/telemetry-plugin/common/output/files';
 import type { OutputOptions } from '@dd/telemetry-plugin/types';
-import { mockLogger } from '@dd/tests/_jest/helpers/mocks';
-import { mockReport } from '@dd/tests/unit/plugins/telemetry/testHelpers';
+import { mockLogger, mockReport } from '@dd/tests/_jest/helpers/mocks';
 import { vol } from 'memfs';
 import path from 'path';
