+++ packages/plugins/telemetry/src/index.test.ts
