+++ packages/tests/jest.config.js
@@ -7,7 +7,7 @@ module.exports = {
     clearMocks: true,
     globalSetup: '<rootDir>/src/_jest/globalSetup.ts',
     preset: 'ts-jest/presets/js-with-ts',
-    roots: ['./src/unit/'],
+    roots: ['<rootDir>/../'],
     setupFilesAfterEnv: ['<rootDir>/src/_jest/setupAfterEnv.ts'],
     testEnvironment: 'node',
     testMatch: ['**/*.test.*'],
