+++ packages/tests/src/_jest/helpers/mocks.ts
@@ -11,11 +11,31 @@ import type {
     Logger,
     LogLevel,
     Options,
+    RepositoryData,
 } from '@dd/core/types';
-import { getSourcemapsConfiguration } from '@dd/tests/unit/plugins/error-tracking/testHelpers';
-import { getTelemetryConfiguration } from '@dd/tests/unit/plugins/telemetry/testHelpers';
+import type {
+    Metadata,
+    MultipartValue,
+    Payload,
+} from '@dd/error-tracking-plugin/sourcemaps/payload';
+import type {
+    SourcemapsOptions,
+    SourcemapsOptionsWithDefaults,
+    Sourcemap,
+} from '@dd/error-tracking-plugin/types';
+import { TrackedFilesMatcher } from '@dd/internal-git-plugin/trackedFilesMatcher';
+import type {
+    Report,
+    Compilation,
+    OptionsDD,
+    OptionsWithTelemetry,
+    OutputOptions,
+    TelemetryOptions,
+    Module,
+} from '@dd/telemetry-plugin/types';
 import { configXpack } from '@dd/tools/bundlers';
-import type { PluginBuild } from 'esbuild';
+import type { PluginBuild, Metafile } from 'esbuild';
+import esbuild from 'esbuild';
 import path from 'path';
 
 import type { BundlerOptionsOverrides, BundlerOverrides } from './types';
@@ -96,7 +116,7 @@ export const getEsbuildMock = (overrides: Partial<PluginBuild> = {}): PluginBuil
     };
 };
 
-export const getMockBuild = (overrides: Partial<BuildReport> = {}): BuildReport => ({
+export const getMockBuildReport = (overrides: Partial<BuildReport> = {}): BuildReport => ({
     errors: [],
     warnings: [],
     logs: [],
@@ -114,10 +134,10 @@ export const getContextMock = (overrides: Partial<GlobalContext> = {}): GlobalCo
     return {
         auth: defaultAuth,
         bundler: {
-            ...getMockBuild().bundler,
+            ...getMockBuildReport().bundler,
             outDir: '/cwd/path',
         },
-        build: getMockBuild(),
+        build: getMockBuildReport(),
         cwd: '/cwd/path',
         env: 'test',
         getLogger: jest.fn(),
@@ -265,3 +285,216 @@ export const filterOutParticularities = (input: File) =>
     !input.filepath.includes('webpack4/buildin') &&
     // Exclude webpack's fake entry point.
     !input.filepath.includes('fixtures/empty.js');
+
+export const getMockPluginBuild = (overrides: Partial<PluginBuild>): PluginBuild => {
+    return {
+        initialOptions: {},
+        esbuild,
+        resolve: jest.fn(),
+        onStart: jest.fn(),
+        onEnd: jest.fn(),
+        onResolve: jest.fn(),
+        onDispose: jest.fn(),
+        onLoad: jest.fn(),
+        ...overrides,
+    };
+};
+
+const mockTapable = { tap: jest.fn() };
+export const mockModule: Module = {
+    name: 'module',
+    userRequest: '',
+    size: 123,
+    loaders: [],
+    chunks: [],
+    _chunks: new Set(),
+    dependencies: [],
+};
+
+export const getMockModule = (overrides: Partial<Module>): Module => ({
+    ...mockModule,
+    ...overrides,
+});
+
+export const mockCompilation: Compilation = {
+    options: {
+        context: '/default/context',
+    },
+    moduleGraph: {
+        getIssuer: () => mockModule,
+        getModule: () => mockModule,
+        issuer: mockModule,
+    },
+    hooks: {
+        buildModule: mockTapable,
+        succeedModule: mockTapable,
+        failedModule: mockTapable,
+        afterOptimizeTree: mockTapable,
+    },
+};
+
+export const getMockCompilation = (overrides: Partial<Compilation>): Compilation => ({
+    options: {
+        ...mockCompilation.options,
+        ...overrides.options,
+    },
+    moduleGraph: {
+        ...mockCompilation.moduleGraph!,
+        ...overrides.moduleGraph!,
+    },
+    hooks: {
+        ...mockCompilation.hooks,
+        ...overrides.hooks,
+    },
+});
+
+export const mockMetaFile: Metafile = {
+    inputs: {
+        module1: {
+            bytes: 1,
+            imports: [],
+        },
+        module2: {
+            bytes: 1,
+            imports: [],
+        },
+    },
+    outputs: {
+        module1: {
+            imports: [],
+            exports: [],
+            inputs: { module2: { bytesInOutput: 0 } },
+            bytes: 0,
+        },
+        module2: {
+            imports: [],
+            exports: [],
+            inputs: { module1: { bytesInOutput: 0 } },
+            bytes: 0,
+        },
+    },
+};
+
+export const mockReport: Report = {
+    timings: {
+        tapables: new Map(),
+        loaders: new Map(),
+        modules: new Map(),
+    },
+};
+
+export const mockOutputOptions: OutputOptions = true;
+export const mockOptionsDD: OptionsDD = {
+    tags: [],
+    prefix: '',
+    timestamp: 1,
+    filters: [],
+};
+export const mockTelemetryOptions: TelemetryOptions = {};
+export const mockOptionsWithTelemetry: OptionsWithTelemetry = {
+    ...defaultPluginOptions,
+    telemetry: mockTelemetryOptions,
+};
+
+export const getTelemetryConfiguration = (
+    overrides: Partial<TelemetryOptions> = {},
+): TelemetryOptions => ({
+    enableTracing: true,
+    endPoint: FAKE_URL,
+    output: true,
+    prefix: 'prefix',
+    tags: ['tag'],
+    timestamp: new Date().getTime(),
+    ...overrides,
+});
+
+export const getMinimalSourcemapsConfiguration = (
+    options: Partial<SourcemapsOptions> = {},
+): SourcemapsOptions => {
+    return {
+        minifiedPathPrefix: '/prefix',
+        releaseVersion: '1.0.0',
+        service: 'error-tracking-build-plugin-sourcemaps',
+        ...options,
+    };
+};
+
+export const getSourcemapsConfiguration = (
+    options: Partial<SourcemapsOptions> = {},
+): SourcemapsOptionsWithDefaults => {
+    return {
+        bailOnError: false,
+        disableGit: false,
+        dryRun: false,
+        maxConcurrency: 10,
+        intakeUrl: INTAKE_URL,
+        minifiedPathPrefix: '/prefix',
+        releaseVersion: '1.0.0',
+        service: 'error-tracking-build-plugin-sourcemaps',
+        ...options,
+    };
+};
+
+export const getSourcemapMock = (options: Partial<Sourcemap> = {}): Sourcemap => {
+    return {
+        minifiedFilePath: '/path/to/minified.min.js',
+        minifiedPathPrefix: '/prefix',
+        minifiedUrl: '/prefix/path/to/minified.js',
+        relativePath: '/path/to/minified.min.js',
+        sourcemapFilePath: '/path/to/sourcemap.js.map',
+        ...options,
+    };
+};
+
+export const getMetadataMock = (options: Partial<Metadata> = {}): Metadata => {
+    return {
+        plugin_version: '1.0.0',
+        project_path: '/path/to/project',
+        service: 'error-tracking-build-plugin-sourcemaps',
+        type: 'js_sourcemap',
+        version: '1.0.0',
+        ...options,
+    };
+};
+
+export const getRepositoryDataMock = (options: Partial<RepositoryData> = {}): RepositoryData => {
+    return {
+        hash: 'hash',
+        remote: 'remote',
+        trackedFilesMatcher: new TrackedFilesMatcher(['/path/to/minified.min.js']),
+        ...options,
+    };
+};
+
+export const getPayloadMock = (
+    options: Partial<Payload> = {},
+    content: [string, MultipartValue][] = [],
+): Payload => {
+    return {
+        content: new Map<string, MultipartValue>([
+            [
+                'source_map',
+                {
+                    type: 'file',
+                    path: '/path/to/sourcemap.js.map',
+                    options: { filename: 'source_map', contentType: 'application/json' },
+                },
+            ],
+            [
+                'minified_file',
+                {
+                    type: 'file',
+                    path: '/path/to/minified.min.js',
+                    options: {
+                        filename: 'minified_file',
+                        contentType: 'application/javascript',
+                    },
+                },
+            ],
+            ...content,
+        ]),
+        errors: [],
+        warnings: [],
+        ...options,
+    };
+};
