--- packages/tests/src/unit/plugins/telemetry/.eslintrc.js
@@ -1,11 +1,0 @@
-module.exports = {
-    overrides: [
-        {
-            files: ['mocks/**/*.*'],
-            rules: {
-                'func-names': 'off',
-                'no-unused-expressions': 'off',
-            },
-        },
-    ],
-};
