+++ packages/tools/src/commands/create-plugin/hooks.test.ts
