+++ packages/tools/src/commands/create-plugin/index.test.ts
