+++ packages/tools/src/commands/integrity/files.test.ts
