+++ packages/tools/src/rollupConfig.test.ts
